Struct corundum::alloc::default::BuddyAlloc

pub struct BuddyAlloc {}

A memory allocator with buddy allocation mechanism

To define a new buddy allocator type as a memory pool, you may use pool!() macro.

Implementations

impl BuddyAlloc

pub fn open_impl(filename: &str) -> Result<Self>

Opens a memory pool file and returns an instance of BuddyAlloc if success. The pool remains open as long as the instance lives.