Module corundum::alloc::open_flags

Open pool flags

Constants

O_1GB
Open Flag: Creates a pool memory file of size 1GB

O_1TB
Open Flag: Creates a pool memory file of size 1TB

O_2GB
Open Flag: Creates a pool memory file of size 2GB

O_2TB
Open Flag: Creates a pool memory file of size 2TB

O_4GB
Open Flag: Creates a pool memory file of size 4GB

O_4TB
Open Flag: Creates a pool memory file of size 4TB

O_8GB
Open Flag: Creates a pool memory file of size 8GB

O_8TB
Open Flag: Creates a pool memory file of size 8TB

O_16GB
Open Flag: Creates a pool memory file of size 16GB

O_16TB
Open Flag: Creates a pool memory file of size 16TB

O_32GB
Open Flag: Creates a pool memory file of size 32GB

O_32TB
Open Flag: Creates a pool memory file of size 32TB

O_64GB
Open Flag: Creates a pool memory file of size 64GB

O_64TB
Open Flag: Creates a pool memory file of size 64TB

O_128GB
Open Flag: Creates a pool memory file of size 128GB

O_256GB
Open Flag: Creates a pool memory file of size 256GB

O_512GB
Open Flag: Creates a pool memory file of size 512GB

O_C
Open Flag: Create the pool memory file

O_CF
Open Flag: Creates and formats a new file

O_CFNE
Open Flag: Creates and formats pool memory file only if it does not exist

O_CNE
Open Flag: Creates pool memory file only if it does not exist

O_F
Open Flag: Formats the pool memory file if file exists, otherwise error