Trait corundum::cell::RootObj

pub trait RootObj<A: MemPool> {
    pub fn init(journal: &Journal<A>) -> Self;
}

Creates a default value of the type

The root type should implement this trait or trait Default to be able to initialize the root object for the first time. Every type implementing Default is already implementing RootObj, by default.

Required methods

pub fn init(journal: &Journal<A>) -> Self

Loading content...