Module corundum::default

The default allocator module

Re-exports

pub use $crate::*;

pub use $crate::alloc::*;

pub use $crate::clone::PClone;

pub use $crate::convert::PFrom;

pub use $crate::str::ToPString;

pub use $crate::stm::transaction;

Modules

parc

prc

Structs

BuddyAlloc
A memory allocator with buddy allocation mechanism

Ptr
A wrapper around a raw persistent pointer that indicates that the possessor of this wrapper owns the referent. Useful for building abstractions like Pbox<T,P>, Vec<T,P>, and String<P>.

RootCell
Root object container

Traits

RootObj
Creates a default value of the type

Type Definitions

Journal
Compact form of Journal <BuddyAlloc>.

PCell
Compact form of PCell <T,BuddyAlloc>.

PMutex
Compact form of PMutex <T,BuddyAlloc>.

PNonNull
Compact form of LogNonNull <T,BuddyAlloc>.

PRef
Compact form of Ref <'b, T, BuddyAlloc>.

PRefCell
Compact form of PRefCell <T,BuddyAlloc>.

PRefMut
Compact form of RefMut <'b, T, BuddyAlloc>.

PString
Compact form of String <BuddyAlloc>.

PVec
Compact form of Vec <T,BuddyAlloc>.

Parc
Compact form of Parc <T,BuddyAlloc>.

Pbox
Compact form of Pbox <T,BuddyAlloc>.

Prc
Compact form of Prc <T,BuddyAlloc>.

VCell
Compact form of VCell <T,BuddyAlloc>.