Struct corundum::prc::Prc

pub struct Prc<T: PSafe + ?Sized, A: MemPool> { /* fields omitted */ }

A single-thread reference-counting persistent pointer. 'Prc' stands for 'Persistent Reference Counted'.

The main aspect of Prc<T> is that its counters are transactional which means that functions pclone, downgrade, and upgrade require a Journal to operate. In other words, you need to wrap them in a transaction.

Prc uses reference counting to manage memory. Although it provides a fast solution for deallocation without scan, cyclic references yield a memory leak. At this point, we have not provided a static solution to detect cyclic references. However, following Rust's partial solution for that, you may use Weak references for reference cycles.

References to data can be strong (using pclone), weak (using downgrade), or volatile weak (using volatile). The first two generate NV-to-NV pointers, while the last on is a V-to-NV pointer. Please see Weak and VWeak for more details on their implementation and safety.

Examples

use corundum::prc::Prc;
use corundum::clone::PClone;
 
P::transaction(|j| {
    let p = Prc::<i32,P>::new(1, j);
 
    // Create a new persistent strong reference
    let s = p.pclone(j);
 
    assert_eq!(*p, *s);
    assert_eq!(2, Prc::strong_count(&p));
    assert_eq!(0, Prc::weak_count(&p));
 
    // Create a new persistent weak reference
    let w = Prc::downgrade(&p, j);
    assert_eq!(2, Prc::strong_count(&p));
    assert_eq!(1, Prc::weak_count(&p));
 
    // Create a new volatile weak reference
    let v = Prc::volatile(&p);
    assert_eq!(2, Prc::strong_count(&p));
    assert_eq!(1, Prc::weak_count(&p));
 
    // Upgrade the persistent weak ref to a strong ref
    let ws = w.upgrade(j).unwrap();
    assert_eq!(3, Prc::strong_count(&p));
    assert_eq!(1, Prc::weak_count(&p));
 
    // Upgrade the volatile weak ref to a strong ref
    let vs = w.upgrade(j).unwrap();
    assert_eq!(4, Prc::strong_count(&p));
    assert_eq!(1, Prc::weak_count(&p));
}).unwrap();

Implementations

impl<T: PSafe, A: MemPool> Prc<T, A>

pub fn new(value: T, journal: &Journal<A>) -> Prc<T, A>

Constructs a new Prc<T>.

It also creates a DropOnFailure log to make sure that if the program crashes, the allocation drops of recovery.

Examples

use corundum::prc::Prc;

P::transaction(|j| {
    let five = Prc::new(5, j);
}).unwrap();

pub fn new_uninit(journal: &Journal<A>) -> Prc<MaybeUninit<T>, A>

Constructs a new Prc with uninitialized contents.

A DropOnFailure log is taken for the allocation.

Examples

use corundum::prc::Prc;

P::transaction(|j| {
    let mut five = Prc::<u32,Heap>::new_uninit(j);

    let five = unsafe {
        // Deferred initialization:
        Prc::get_mut_unchecked(&mut five).as_mut_ptr().write(5);

        five.assume_init()
    };

    assert_eq!(*five, 5)
}).unwrap();

pub fn new_zeroed(journal: &Journal<A>) -> Prc<MaybeUninit<T>, A>

Constructs a new Prc with uninitialized contents, with the memory being filled with 0 bytes.

See MaybeUninit::zeroed for examples of correct and incorrect usage of this method.

Examples

use corundum::prc::Prc;

P::transaction(|j| {
    let zero = Prc::<i32,P>::new_zeroed(j);
    let zero = unsafe { zero.assume_init() };

    assert_eq!(*zero, 0)
}).unwrap();

pub fn from(p: Prc<T, A>) -> Self

Owns contents of p without cloning, leaving p untouched

impl<T: PSafe, A: MemPool> Prc<MaybeUninit<T>, A>

pub unsafe fn assume_init(self) -> Prc<T, A>

Converts to Rc<T>.

Safety

As with MaybeUninit::assume_init, it is up to the caller to guarantee that the inner value really is in an initialized state. Calling this when the content is not yet fully initialized causes immediate undefined behavior.

Examples

use corundum::prc::Prc;

corundum::transaction(|j| {
    let mut five = Prc::<u32,P>::new_uninit(j);

    let five = unsafe {
        // Deferred initialization:
        Prc::get_mut_unchecked(&mut five).as_mut_ptr().write(5);

        five.assume_init()
    };

    assert_eq!(*five, 5);
}).unwrap();

impl<T: PSafe, A: MemPool> Prc<MaybeUninit<T>, A>

pub fn get_mut(this: &mut Self) -> Option<&mut MaybeUninit<T>>

Returns a mutable reference into the given Prc, if there are no other Prc or Weak pointers to the same allocation.

Returns None otherwise, because it is not safe to mutate a shared value. It only works for Prc<MaybeUninit<T>> to be able to defer the initialization.

Examples

use corundum::prc::Prc;

P::transaction(|j| {
    let mut five = Prc::<u32,P>::new_uninit(j);

    let five = unsafe {
        // Deferred initialization:
        Prc::get_mut(&mut five).unwrap().as_mut_ptr().write(5);

        five.assume_init()
    };

    assert_eq!(*five, 5)
}).unwrap();

pub unsafe fn get_mut_unchecked(this: &mut Self) -> &mut MaybeUninit<T>

Returns a mutable reference into the given Prc, without any check.

It only works for Prc<MaybeUninit<T>> to be able to defer the initialization.

Safety

Any other Prc or Weak pointers to the same allocation must not be dereferenced for the duration of the returned borrow. This is trivially the case if no such pointers exist, for example immediately after Rc::new.

Examples

use corundum::prc::Prc;

P::transaction(|j| {
    let mut five = Prc::<i32,P>::new_uninit(j);

    let five = unsafe {
        // Deferred initialization:
        Prc::get_mut_unchecked(&mut five).as_mut_ptr().write(5);

        five.assume_init()
    };

    assert_eq!(*five, 5);
}).unwrap();

impl<T: PSafe + ?Sized, A: MemPool> Prc<T, A>

pub fn downgrade(this: &Self, journal: &Journal<A>) -> Weak<T, A>

Creates a new Weak pointer to this allocation.

Examples

use corundum::prc::Prc;

P::transaction(|j| {
    let five = Prc::new(5, j);
    let _weak_five = Prc::downgrade(&five, j);
}).unwrap()

pub fn volatile(this: &Self) -> VWeak<T, A>

Creates a new VWeak pointer to this allocation.

pub fn weak_count(this: &Self) -> usize

Gets the number of Weak pointers to this allocation.

Examples

use corundum::prc::Prc;

P::transaction(|j| {
    let five = Prc::new(5, j);

    let _weak_five = Prc::downgrade(&five, j);
    assert_eq!(1, Prc::weak_count(&five));
}).unwrap()

pub fn strong_count(this: &Self) -> usize

Gets the number of Weak pointers to this allocation.

Examples

use corundum::prc::Prc;
use corundum::clone::PClone;

P::transaction(|j| {
    let five = Prc::new(5, j);
    let _also_five = Prc::pclone(&five, j);
    assert_eq!(2, Prc::strong_count(&five));
}).unwrap();

pub fn ptr_eq(this: &Self, other: &Self) -> bool

Returns true if the two Prcs point to the same allocation (in a vein similar to ptr::eq).

Examples

use corundum::prc::Prc;
use corundum::clone::PClone;

P::transaction(|j| {
    let five = Prc::new(5, j);
    let same_five = Prc::pclone(&five, j);
    let other_five = Prc::new(5, j);

    assert!(Prc::ptr_eq(&five, &same_five));
    assert!(!Prc::ptr_eq(&five, &other_five));
}).unwrap();