Module corundum::stm

Software transactional memory APIs

Modules

pspd
Persistent Scratchpad Memory

vspd
Volatile Scratchpad Memory

Structs

Chaperon
A third-party observer for multi-pool transactions

Journal
A Journal object to be used for writing logs onto

Log
The Log type for pool A

Enums

LogEnum
Log Types

Notifier
A data-log notification type

Constants

JOURNAL_COMMITTED
Determines that the changes are committed

Traits

Logger
A generic trait for taking a log of any type

Functions

transaction
Atomically executes commands