Trait corundum::stm::Logger

pub trait Logger<A: MemPool> {
    unsafe fn take_log(
        &self, 
        journal: &Journal<A>, 
        notifier: Notifier<A>
    ) -> Ptr<Log<A>, A>;
}

A generic trait for taking a log of any type

Required methods

unsafe fn take_log(
&self,
journal: &Journal<A>,
notifier: Notifier<A>
) -> Ptr<Log<A>, A>

Takes a log of self and update the log flag if specified in notifier

Loading content...