Trait corundum::PSafe

pub unsafe auto trait PSafe { }

It marks the implementing type to be free of pointers to the volatile heap, and persistence safe.

Also, every type that allows interior mutability is not safe in persistence terms, because there might be no log of the value. Atomic types are persistence safe, even though they provide interior mutability.