Trait corundum::PSend

pub unsafe auto trait PSend { }

Safe to be sent to another thread

This marker is used to allow Parc to be sent to another thread only if it is wrapped in a VWeak. The Parc is not Send to prevent escaping a newly allocated instance of it from a transaction.