Trait corundum::TxInSafe

pub unsafe auto trait TxInSafe { }

It is equal to UnwindSafe, but is used to ensure doubly that mutable references cannot go inside a transaction.

Safety

The user can safely specify a type as UnwindSafe, but TxInSafe is unsafe to implement. This warns the programmer that the non-existence of orphans is not guaranteed anymore.